Top-Bar
Made the TopBar for the dashboard
