+++ src/styles/TopBar.css
@@ -1,0 +1,23 @@
+.topbar{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 20px;
+    background-color: #474747;
+    color: white;
+}
+.topbar h1 {
+    font-size: 24px;
+    margin: 0;
+    font-family: 'Press Start 2P', cursive;
+}
+.logout-btn{
+    background-color: rgb(255, 32, 32);
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    font-size: 16px;
+    cursor: pointer;
+    font-family: 'Press Start 2P', cursive;
+    border-radius: 5px;
+}
