+++ src/styles/dashboard.css
@@ -1,15 +1,13 @@
-.dashboard-container {
-    background-image: url('/backgrounds/dashboard-bg.png');
-    background-size: cover;
-    background-position: center;
-    min-height: 100vh;
-    color: white;
-    font-family: 'Press Start 2P', cursive;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 2rem;
-    text-align: center;
-  }
+.dashboard-content {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 2rem;
+  text-align: center;
+  min-height: 90vh;
+  font-family: 'Press Start 2P', cursive;
+  color: white;
+}
+
